arrange files in folders
