+++ bin/main.rb
@@ -1,0 +1,3 @@
+require_relative './item'
+
+item = Item.new.get_items
