+++ lib/receipt.rb
@@ -1,0 +1,11 @@
+class Receipt
+  def print_tax(tax_sum)
+    print "Sales Taxes: "
+    puts tax_sum
+  end
+
+  def print_total_price(total_price)
+    print "Total: "
+    puts total_price
+  end
+end
