--- product.rb
@@ -1,3 +1,0 @@
-class Product
-  
-end
