--- receipt.rb
@@ -1,9 +1,0 @@
-class Receipt
-  def initialize()
-    puts "Printing Receipt..."
-  end
-
-  def print
-    "Cant process this yet..."
-  end
-end
