--- tax.rb
@@ -1,3 +1,0 @@
-class Tax
-  
-end
